requirements.txt
numpy 
opencv-python 
scikit-image
Pillow
pandas
matplotlib

tqdm 
setproctitle


tensorflow-addons==0.10.0
tensorboard-plugin-profile
tensorboard

wandb
jupyterlab 



